Add form for resetting password
This form uses a hidden_field_tag for the email. As the user should not
submit their email, but we need the value to find the user because the
controller actions expect it. When submit the field we lose the @user
object, therefor we have to submit it in a hidden_field.