Add ability to sort cheats by category for a given Cheatsheet
Previously the show cheatsheet view had far too much logic that caused
numerous issues and was not maintainable nor extendable.

This commit adds a simple method to the cheatsheet class which returns
a hash of cheats with keys for each category. The corresponding value is
a collection of all the cheats of that category.

Solution from:
http://vaidehijoshi.github.io/blog/2015/06/09/refactoring-to-reveal-rails-group-by/